Move ChannelInput to a better spot
